Revert "kinoite: Remove initial-setup"
Removing initial-setup-gui also removes a lot of dependencies (dmraid,
bcache-tools, NetworkManager-team, mobile-broadband-provider-info,
ndctl) and this may create issues for users if they rely on them.

As we are not in a hurry to remove that from Fedora 37 (as it won't fix
any bug), we'll remove it by default in Fedora 38 only as we'll have
time to test that nothing breaks there.

This reverts commit 73a441aab3b1678c7152dd61d8ee8dfa421b1d9f.